[Makefile] Added Makefile

+++ controller/studentController.go
@@ -12,13 +12,12 @@ import (
 func GetStudents(context *gin.Context) {
 	students := database.GetAllStudents()
 	fmt.Println(students)
-	sendJson(context.Writer, 200, toJson(students))
+	sendJson(context.Writer, http.StatusOK, toJson(students))
 }
 
 func AddStudent(context *gin.Context) {
 	student := getStudentDetails(context)
 	if student == nil {
-		context.Status(400)
 		return
 	}
 	student = database.Insert(*student)
@@ -26,14 +25,35 @@ func AddStudent(context *gin.Context) {
 }
 
 func DeleteStudent(context *gin.Context) {
-	var rollNum, _ = context.GetQuery("RollNumber")
+	var rollNum, has = context.GetQuery("RollNumber")
+	if !has || rollNum == "" {
+		sendJson(context.Writer, http.StatusBadRequest, toJson(gin.H{
+			"error": "Invalid query. The query must contain 'RollNumber' which needed to be deleted",
+		}))
+		return
+	}
 	_, student := database.DeleteByRollNumber(rollNum)
+	if student == nil {
+		sendJson(context.Writer, http.StatusNotFound, toJson(gin.H{
+			"error": "unable to find the record for the given RollNumber '" + rollNum + "'",
+		}))
+		return
+	}
 	sendJson(context.Writer, http.StatusOK, toJson(student))
 }
 
 func UpdateStudent(context *gin.Context) {
 	student := getStudentDetails(context)
+	if student == nil {
+		return
+	}
 	student = database.UpdateByRollNumber(student.RollNumber, *student)
+	if student == nil {
+		sendJson(context.Writer, http.StatusNotFound, toJson(gin.H{
+			"error": "unable to find the record for the given RollNumber '" + student.RollNumber + "'",
+		}))
+		return
+	}
 	sendJson(context.Writer, http.StatusOK, toJson(student))
 }
 
@@ -42,6 +62,9 @@ func getStudentDetails(context *gin.Context) *models.Student {
 	err := context.BindJSON(&student)
 	if err != nil {
 		fmt.Println(err)
+		sendJson(context.Writer, http.StatusBadRequest, toJson(gin.H{
+			"error": "Please send the body for the request, which must contains 'Name' and 'RollNumber'",
+		}))
 		return nil
 	}
 	return &student
@@ -54,5 +77,6 @@ func toJson(obj any) []byte {
 
 func sendJson(writer gin.ResponseWriter, code int, json []byte) {
 	writer.WriteHeader(code)
+	writer.Header().Set("Content-Type", "application/json")
 	writer.Write(json)
 }
